functions/requirements.txt
firebase-admin
google-cloud-firestore
google-generativeai
functions-framework
firebase-functions>=0.2.0
